Remove default value for code_signing_identity, code_sign_style, and provisioning_profile properties
xcodebuild (this gem, not the Xcode CLI) failed to work with Xcode 14 with “manual” code signing. Typically, the following would be passed to xcodebuild (the Xcode CLI): CODE_SIGN_STYLE=…, CODE_SIGN_IDENTITY=…, and PROVISIONING_PROFILE_SPECIFIER=….

Build settings set this way apply to all targets including any pod dependency targets, even if the build settings aren’t applicable to the target. In particular, resource bundles are not code signed. Xcode 14 considers settings signing-related settings on a resource bundle to be an error. Prior versions of Xcode did not treat this as an error.

This change requires that the appropriate code signing setting now be set in the Xcode project file.